Move TestPromptTemplateSyntax to test_prompt_template.py
